requirements.txt
asgiref==3.4.1
Babel==2.9.1
beautifulsoup4==4.9.3
Django==3.2.6
django-money==2.0.1
django-multiupload==0.5.2
importlib-metadata==3.10.1
Pillow==8.3.1
psycopg2==2.9.1
py-moneyed==1.2
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
zipp==3.5.0
